Refactor Menu
